+++ packages/rpc/src/index.ts
@@ -1,2 +1,3 @@
+export * from "./fetch-with-deadline";
 export * from "./client";
 export * from "./indexer";